% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_enrichment.R
\name{compute_enrichment}
\alias{compute_enrichment}
\title{Compute enrichment}
\usage{
compute_enrichment(dbs, sobj, res_search, output_path)
}
\arguments{
\item{dbs}{A character specifying the GO category}

\item{sobj}{A Seurat object}

\item{res_search}{A vector of type double that specifies the resolution parameters}

\item{output_path}{A character denoting the output path}
}
\description{
Compute an enrichment analysis of each community using their differentially expressed genes
at each resolution parameter.
}
