% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_modularity.R
\name{compute_modularity}
\alias{compute_modularity}
\title{Compute modularity}
\usage{
compute_modularity(sobj, res_search, output_path)
}
\arguments{
\item{sobj}{A Seurat object}

\item{res_search}{A sequence of louvain resolution parameters}

\item{output_path}{Path of the file containing the louvain cluster modularity scores at each resolution}
}
\value{
cl.setup A list containing the Seurat object and modularity dataframe
}
\description{
Compute modularity
}
