% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_compute_multi_res_search.R
\name{compute_multi_res_search}
\alias{compute_multi_res_search}
\title{Compute multi-resolution search}
\usage{
compute_multi_res_search(sobj, res_search, output_path)
}
\arguments{
\item{sobj}{A Seurat object}

\item{res_search}{A vector of type double that specifies the resolution parameters}

\item{output_path}{A character denoting the output path}
}
\description{
Compute clustering for all resolution parameters.
}
