% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_silhouette_scores.R
\name{wrangle_silhouette_scores}
\alias{wrangle_silhouette_scores}
\title{Wrangle silhouette scores}
\usage{
wrangle_silhouette_scores(sil_list)
}
\arguments{
\item{sil_list}{A list of silhouette score summary statistics for each resolution's communities}
}
\value{
sil_df
}
\description{
Wrangle silhouette scores
}
