% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_modularity.R
\name{compute_modularity}
\alias{compute_modularity}
\title{Compute modularity}
\usage{
compute_modularity(sobj, res_search, output_path)
}
\arguments{
\item{sobj}{A Seurat object}

\item{res_search}{A vector of type double that specifies the resolution parameters}

\item{output_path}{A character denoting the output path}
}
\description{
Compute clustering for all resolution parameters.
}
