% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_enrichment.R
\name{select_geneset_markers}
\alias{select_geneset_markers}
\title{Select geneset markers
Default is the top 5 genes}
\usage{
select_geneset_markers(markers_list)
}
\value{

}
\description{
Select geneset markers
Default is the top 5 genes
}
