% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_enrichment.R
\name{select_geneset_markers}
\alias{select_geneset_markers}
\title{Select geneset markers}
\usage{
select_geneset_markers(markers_list)
}
\arguments{
\item{markers_list}{A list of type character specifying the differentially expressed genes at each resolution and community}
}
\description{
Create a dataframe of the top differentially expressed genes for communities across all resolution parameters.
The default is set to return the top 5 genes.
}
